// packages/client-airtable/src/enums.ts
export enum UpdateType {
    /** Will update the included fields and clear all unincluded fields. */
    FULL = 'PUT',

    /** Will only update the included fields. Fields not included will be unchanged. */
    PARTIAL = 'PATCH',
}